move code in order to support layers
